code/pyproject.toml
[project]
name = "sqlagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.4.1",
    "dspy-ai>=3.0.4",
    "ipywidgets>=8.1.8",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "openai>=2.8.0",
    "pandas>=2.3.3",
    "pip>=25.3",
    "psycopg-binary>=3.2.13",
    "psycopg>=3.2.13",
    "pydot>=4.0.1",
    "pygraphviz>=1.14",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "tqdm>=4.67.1",
]
